services/background/requirements.txt
pillow
numpy
torch
torchvision
diffusers
transformers
accelerate
safetensors
opencv-python
pyyaml
httpx
tenacity
python-json-logger